+++ lib/CMakeLists.txt
@@ -5,3 +5,4 @@ add_subdirectory(libc)
 endif()
 add_subdirectory_if_kconfig(ring_buffer)
 add_subdirectory_if_kconfig(base64)
+add_subdirectory(mempool)
