+++ lib/mempool/CMakeLists.txt
@@ -1,0 +1,1 @@
+zephyr_sources(mempool.c)
